bff-venta/requirements.txt
Flask>=2.3,<4
boto3>=1.28
python-dotenv>=1.0
gunicorn>=21.2